add default HF modal behaviour & fix blank page
