+++ src/components/GridLayoutPage/GridLayoutPage.container.js
@@ -1,7 +1,6 @@
 import { connect } from 'react-redux'
 
 import UIActions from '../../redux/actions/ui'
-import { MARKET_PAGE } from '../../redux/constants/ui'
 import {
   getLayouts, getActiveMarket, getActiveExchange,
 } from '../../redux/selectors/ui'
@@ -12,8 +11,6 @@ const mapStateToProps = (state = {}) => ({
   layouts: getLayouts(state),
   activeMarket: getActiveMarket(state),
   exID: getActiveExchange(state),
-  firstLogin: state.ui.firstLogin,
-  isGuideActive: state.ui[`${MARKET_PAGE}_GUIDE_ACTIVE`],
 })
 
 const mapDispatchToProps = dispatch => ({
@@ -27,10 +24,6 @@ const mapDispatchToProps = dispatch => ({
 
   deleteLayout: (id) => {
     dispatch(UIActions.deleteLayout(id))
-  },
-
-  finishGuide() {
-    dispatch(UIActions.finishGuide(MARKET_PAGE))
   },
 })
