+++ src/pages/MarketData/MarketData.container.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux'
 import {
   getLayouts, getActiveMarket, getActiveExchange,
 } from '../../redux/selectors/ui'
+import { MARKET_PAGE } from '../../redux/constants/ui'
 import UIActions from '../../redux/actions/ui'
 
 import MarketData from './MarketData'
@@ -10,6 +11,9 @@ const mapStateToProps = (state = {}) => ({
   layouts: getLayouts(state),
   activeMarket: getActiveMarket(state),
   exID: getActiveExchange(state),
+  firstLogin: state.ui.firstLogin,
+  isGuideActive: state.ui[`${MARKET_PAGE}_GUIDE_ACTIVE`],
+
 })
 
 const mapDispatchToProps = dispatch => ({
@@ -23,6 +27,10 @@ const mapDispatchToProps = dispatch => ({
 
   deleteLayout: (id) => {
     dispatch(UIActions.deleteLayout(id))
+  },
+
+  finishGuide() {
+    dispatch(UIActions.finishGuide(MARKET_PAGE))
   },
 })
