+++ src/pages/MarketData/MarketData.props.js
@@ -1,0 +1,9 @@
+import PropTypes from 'prop-types'
+
+export const propTypes = {
+  finishGuide: PropTypes.func,
+  isGuideActive: PropTypes.bool,
+  irstLogin: PropTypes.bool,
+}
+
+export const defaultProps = {}
