+++ src/pages/Trading/Trading.js
@@ -67,8 +67,8 @@ export default class Trading extends React.PureComponent {
     const { finishGuide } = this.props
     const { status } = data
     const finishedStatuses = [STATUS.FINISHED, STATUS.SKIPPED]
-
-    if (finishedStatuses.includes(status)) {
+    const CLOSE = 'close'
+    if (finishedStatuses.includes(status) || data.action === CLOSE) {
       finishGuide()
     }
   }
@@ -76,6 +76,7 @@ export default class Trading extends React.PureComponent {
   render() {
     const { activeMarket, firstLogin, isGuideActive } = this.props
     const { steps } = this.state
+    console.log('trading page guide')
     const commonComponentProps = {
       layoutID: LAYOUT_ID,
       moveable: true,
